AB#2659 -- Merge all auth routes into one
