+++ frontend/.env.example
@@ -21,7 +21,7 @@ AUTH_JWT_PRIVATE_KEY=MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQCwaAVE/rA
 
 # RAOIDC base URL -- used as a base URL for all RAOIDC calls
 # (required; use the example below)
-AUTH_RAOIDC_BASE_URL=https://auth.example.com/oauth
+AUTH_RAOIDC_BASE_URL=http://localhost:3000/auth
 # RAOIDC client ID -- the OAuth client id used for all RAOIDC calls
 # (required; use the example below)
 AUTH_RAOIDC_CLIENT_ID=CDCP
